Refactor: parse.js -> parse-util.js, error.js
